requirements.txt
# Core Dependencies
apache-airflow==3.0.6
pandas==2.3.2
numpy==2.0.1
requests==2.32.5
openpyxl==3.1.5
xlsxwriter==3.2.9

# ETL Script Dependencies
fire==0.7.1
python-dotenv==1.1.1

# Email Dependencies
aiosmtplib==4.0.2

# Progress and Utilities
tqdm==4.67.1
python-dateutil==2.8.2
pytz==2023.3

# Optional: Machine Learning (for clair-py310 compatibility)
tensorflow==2.19.1
keras==3.11.3
scikit-learn==1.3.0
scipy==1.15.3

# Optional: Data Visualization
matplotlib==3.10.6
seaborn==0.12.2

# Optional: Web Framework (if needed)
fastapi==0.117.1
uvicorn==0.37.0

# Development and Testing
pytest==7.4.0
black==23.0.0
flake8==6.0.0